separate delve simulation location
